The JupyterLab App Launcher config file is now auto-generated on terminal login
This is done via a script in .bash_profile and required converting the existing JupyterLab App Launcher config file into a template. The use case that this implements is that a URL to Bandage can be auto-generated and added to the config before JupyterLab starts, meaning the link will always be valid even though the external IP address of the VM changes every time the VM restarts.